Clean up SDK runtime workflow modules
